Cypress command for Okta SSO.
