+++ package.json
@@ -7,7 +7,7 @@
     "cypress": "cypress open"
   },
   "dependencies": {
-    "cypress": "^3.4.0"
+    "cypress": "^3.4.1"
   },
   "devDependencies": {
     "eslint-plugin-cypress": "^2.7.0"
